WIP: Add departed state for members
This PR adds a fourth state for members, 'Departed'. A departed member
is no longer counted as valid for calculating the total population
needed for a quorum.

When a member is marked `Confirmed`, we record the time from the expired
thread. When the expired thread runs, it checks for members whose
departure time has come; and if it has, it sets their health to
`Departed`, removes them from the departure timer, and shares the
rumor.

This commit sets the departure timer at 72 hours, a particularly
conservative number. If an administrator does nothing to bring member
back to life in that timeframe, it will be marked departed
automatically.

TODO:

* Have a command line tool in `hab-butterfly` to manually mark a node as
  Departed
* Automatically depart members if the service group has `Confirmed`
  members, and new members have been added.

Signed-off-by: Adam Jacob <adam@chef.io>